creating a python flask RESTful API
A RESTful API is an application program interface that uses HTTP requests to GET, PUT, POST and DELETE data. REST (REpresentational State Transfer) is an architectural style and approach to communications in web services development. RESTful API explicitly takes advantage of HTTP methodologies like GET to retrieve a resource, PUT to change the state of or update a resource which can be an object file or block, POST to create that resource and DELETE to remove it. I will be creating s set of API endpoints.